refactor: modularized twitch mod
* Make Twitch module less dependant on provided HTTP Client
* s/vod/video
